Move store code to src/store/
