+++ src/ast/index.ts
@@ -1,3 +1,4 @@
 export { getFieldArguments, normalizeVariables } from './variables';
+export { SchemaPredicates } from './schemaPredicates';
 export * from './traversal';
 export * from './node';
