--- src/helpers/dict.ts
@@ -1,1 +1,0 @@
-export const makeDict = (): any => Object.create(null);
