+++ src/operations/shared.ts
@@ -1,11 +1,11 @@
 import { FieldNode, InlineFragmentNode, FragmentDefinitionNode } from 'graphql';
-import { Fragments, Variables, SelectionSet, Scalar } from '../types';
-import { Store } from '../store';
-import { keyOfField } from '../helpers';
+
 import { warn, pushDebugNode } from '../helpers/help';
-import { SchemaPredicates } from '../ast/schemaPredicates';
+import { Store, keyOfField } from '../store';
+import { Fragments, Variables, SelectionSet, Scalar } from '../types';
 
 import {
+  SchemaPredicates,
   getTypeCondition,
   getFieldArguments,
   shouldInclude,
