+++ src/store/index.ts
@@ -1,0 +1,3 @@
+export { makeDict } from './data';
+export * from './keys';
+export * from './store';
